+++ app/components/signup/signup-state.js
@@ -1,17 +1,35 @@
 import { observable, action } from 'mobx';
 import { mainState, uiState, loginState } from '../states';
 import RoutedState from '../routes/routed-state';
-import { User, crypto } from '../../lib/icebear';
+import {
+    User,
+    crypto,
+    saveAccountKeyBackup,
+    config,
+    validation
+} from '../../lib/icebear';
+import { tx } from '../utils/translator';
+import { when } from '../../../node_modules/mobx/lib/mobx';
+
+const { validators } = validation;
+const { suggestUsername } = validators;
 
 class SignupState extends RoutedState {
     @observable username = '';
     @observable email = '';
     @observable firstName = '';
     @observable lastName = '';
+    @observable passphrase = '';
     @observable pin = '';
     @observable _current = 0;
-    get current() { return this._current; }
-    set current(i) { uiState.hideAll().then(() => { this._current = i; }); }
+    get current() {
+        return this._current;
+    }
+    set current(i) {
+        uiState.hideAll().then(() => {
+            this._current = i;
+        });
+    }
     _prefix = 'signup';
     avatarBuffers = null;
     @observable avatarData = null;
@@ -20,12 +38,17 @@ class SignupState extends RoutedState {
     @observable specialty = '';
     @observable role = '';
     @observable medicalId = '';
+    @observable usernameSuggestions = [];
+    @observable subscribeToPromoEmails = false;
 
-    get isFirst() { return this.current === 0; }
+    get isFirst() {
+        return this.current === 0;
+    }
 
     transition = () => this.routes.app.signupStep1();
 
-    @action.bound exit() {
+    @action.bound
+    exit() {
         this.username = '';
         this.email = '';
         this.firstName = '';
@@ -35,45 +58,108 @@ class SignupState extends RoutedState {
         this.country = '';
         this.specialty = '';
         this.role = '';
+        this.keyBackedUp = false;
         this.current = 0;
-        this.routes.app.loginStart();
+        this.usernameSuggestions.clear();
+        this.routes.app.loginWelcome();
 
         // hook for whitelabel signup state to reset itself
         if (this.onExitHandler) this.onExitHandler();
     }
 
-    @action reset() { this.current = 0; }
+    @action
+    reset() {
+        this.current = 0;
+    }
 
     generatePassphrase = () => crypto.keys.getRandomAccountKeyHex();
 
-    @action async next() {
+    @action.bound
+    async next() {
         if (!this.passphrase) this.passphrase = await this.generatePassphrase();
-        if (this.keyBackedUp && !User.current) await this.finishAccountCreation();
         this.current++;
     }
 
-    @action prev() { (this.current > 0) ? this.current-- : this.exit(); }
-
-    @action async finishSignUp() {
-        return mainState.activateAndTransition(User.current)
-            .catch((e) => {
-                console.log(e);
-                User.current = null;
-                this.reset();
-            });
+    @action.bound
+    prev() {
+        this.current > 0 ? this.current-- : this.exit();
     }
 
-    // After account is created, user goes to Contact Sync rather than main route
-    @action async finishAccountCreation() {
+    @action.bound
+    async suggestUsernames() {
+        try {
+            this.usernameSuggestions = await suggestUsername(
+                this.firstName,
+                this.lastName
+            );
+        } catch (e) {
+            console.error(e);
+        }
+    }
+
+    @action
+    async finishSignUp() {
+        return mainState.activateAndTransition(User.current).catch(e => {
+            console.log(e);
+            User.current = null;
+            this.reset();
+        });
+    }
+
+    get backupFileName() {
+        return `${this.username}-${tx('title_appName')}.pdf`;
+    }
+
+    @action.bound
+    async saveAccountKey() {
+        const {
+            username,
+            firstName,
+            lastName,
+            passphrase,
+            backupFileName
+        } = this;
+        const fileSavePath = config.FileStream.getTempCachePath(backupFileName);
+        await saveAccountKeyBackup(
+            fileSavePath,
+            `${firstName} ${lastName}`,
+            username,
+            passphrase
+        );
+        this.isInProgress = true;
+        try {
+            await config.FileStream.launchViewer(fileSavePath);
+            this.keyBackedUp = true;
+        } catch (e) {
+            console.error(e);
+        }
+        this.isInProgress = false;
+    }
+
+    @action
+    async finishAccountCreation() {
         this.isInProgress = true;
         const user = new User();
         User.current = user;
-        const { username, email, firstName, lastName, passphrase, avatarBuffers,
-            keyBackedUp, country, specialty, role, medicalId } = this;
+        const {
+            username,
+            email,
+            firstName,
+            lastName,
+            passphrase,
+            avatarBuffers,
+            keyBackedUp,
+            subscribeToPromoEmails,
+            country,
+            specialty,
+            role,
+            medicalId
+        } = this;
         const localeCode = uiState.locale;
         user.username = username;
         user.email = email;
-        user.passphrase = __DEV__ && process.env.PEERIO_QUICK_SIGNUP ? 'icebear' : passphrase;
+        user.passphrase =
+            __DEV__ && process.env.PEERIO_QUICK_SIGNUP ? 'icebear' : passphrase;
         user.firstName = firstName;
         user.lastName = lastName;
         user.localeCode = localeCode;
@@ -89,14 +175,29 @@ class SignupState extends RoutedState {
                 mcrAHPRA: medicalId
             };
         }
-        return user.createAccountAndLogin()
+        return user
+            .createAccountAndLogin()
             .then(() => loginState.enableAutomaticLogin(user))
             .then(() => mainState.saveUser())
             .then(() => keyBackedUp && User.current.setAccountKeyBackedUp())
             .then(() => avatarBuffers && User.current.saveAvatar(avatarBuffers))
-            .finally(() => { this.isInProgress = false; });
+            .then(() => {
+                // TODO: replace with icebear version after it's merged
+                const { settings } = User.current;
+                when(
+                    () => !settings.loading,
+                    () => {
+                        settings.subscribeToPromoEmails = subscribeToPromoEmails;
+                        User.current.saveSettings();
+                    }
+                );
+            })
+            .finally(() => {
+                this.isInProgress = false;
+            });
     }
 }
 
 const signupState = new SignupState();
+global.signupState = signupState;
 export default signupState;