+++ app/styles/vars.js
@@ -100,7 +100,9 @@ const vars = {
     toggleLineActive: '#B6D3FF',
     toggleLineInactive: '#CFCFCF',
     imageInnerContainerHeight: 140,
-    peerioFontFamily: 'Open Sans',
+    peerioFontFamily: 'OpenSans',
+    peerioSerifFontFamily: 'SourceSerifPro',
+    peerioCourierFontFamily: 'Courier',
     chatUnreadIndicatorBg: 'rgba(255, 255, 255, 0.95)',
     chatUnreadIndicatorWidth: 70,
     chatUnreadIndicatorHeight: 32,
@@ -110,6 +112,7 @@ const vars = {
     popupHorizontalMargin: scaleDim(16),
     topDrawerHeight: 192,
     tfaInputWidth: scaleDim(124),
+    signupButtonWidth: 82,
     scrollOffset: 500, // arbitrary large offset
 
     circle: 10,
@@ -177,6 +180,7 @@ const vars = {
     footerMarginX: 24,
     statusBarHeight,
     layoutPaddingTop,
+    welcomeHeaderHeight: 80 + statusBarHeight,
     headerHeight: 56,
     headerSpacing: 56 + layoutPaddingTop,
     headerIconMargin: 16,
@@ -197,7 +201,7 @@ const vars = {
     chatListItemDMHeight: 61,
     removeButtonHeight: 44,
     warningHeight: 70,
-    roundedButtonWidth: 134,
+    roundedButtonWidth: 155,
     wideRoundedButtonWidth: 190,
     menuWidthRatio: 0.8,
     animationDuration: 200,
@@ -229,6 +233,7 @@ const vars = {
     modalPaddingHorizontal: 40,
     wizardPadding: 36,
     height80: height * 0.8,
+    topCircleSizeSmall: 52,
 
     font: {
         size: {
@@ -292,7 +297,8 @@ const vars = {
             midixx: scaleDim(34),
             midi2x: scaleDim(36),
             maxi: scaleDim(40),
-            maxi2x: scaleDim(48)
+            maxi2x: scaleDim(44),
+            maxi3x: scaleDim(48)
         },
         huge: {
             mini: scaleDim(50),
@@ -302,7 +308,8 @@ const vars = {
             midi2x: scaleDim(64),
             midi3x: scaleDim(68),
             maxi: scaleDim(74),
-            maxi2x: scaleDim(75)
+            maxi2x: scaleDim(75),
+            maxi3x: scaleDim(80)
         }
     },
     loadingTimeout: 15000,
@@ -317,9 +324,9 @@ const vars = {
     button: {
         touchableHeight: 48,
         buttonHeight: 36,
-        minWidth: 72,
+        minWidth: 96,
         borderRadius: 24,
-        paddingHorizontal: scaleDim(16),
+        paddingHorizontal: scaleDim(12),
         marginVertical: scaleDim(8),
         fontSize: scaleDim(14)
     }